recreate instance buffer when needed
